Move docker row schemes to schemes folder
